# docs/swagger.yaml
definitions:
...
  errors.HttpError:
    properties:
      code:
        type: integer
      message:
        type: string
      stack_trace:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /stream/{objectPath}:
    get:
      consumes:
      - application/json
      description: Streams video content from MinIO with support for range requests
      parameters:
      - description: Object path in the bucket
        in: path
        name: objectPath
        required: true
        type: string
      - description: Range header for partial content requests
        in: header
        name: Range
        type: string
      - description: API Key for authentication
        in: header
        name: X-API-KEY
        required: true
        type: string
      produces:
      - video/mp4
      responses:
        "200":
          description: Full video content
          schema:
            type: file
        "206":
          description: Partial video content
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.HttpError'
      summary: Stream video content
      tags:
      - Stream
...
swagger: "2.0"